backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic==2.10.3
python-dotenv==1.0.1
pymongo==4.9.1
reportlab==4.2.5
resend==2.19.0